Create DriverLogger and move access list metrics into it
