+++ crates/solver/src/lib.rs
@@ -2,6 +2,7 @@ mod analytics;
 pub mod arguments;
 mod auction_preprocessing;
 pub mod driver;
+pub mod driver_logger;
 pub mod encoding;
 pub mod in_flight_orders;
 pub mod interactions;
